Separate `impl` and `solve` tests
Move all mock definitions to `mock_integration.py' file
